Move field logic to hooks
