+++ lib/field/field.tsx
@@ -1,4 +1,3 @@
-import { ZodError } from "zod";
 import {
   ForwardedRef,
   forwardRef,
@@ -6,19 +5,17 @@ import {
   useCallback,
   useContext,
   useImperativeHandle,
-  useLayoutEffect,
   useMemo,
   useRef,
-  useState,
 } from "react";
 import { FieldInstanceProps, FieldInstance } from "./types";
 import { FormContext } from "../form/context";
-import { getValidationError, stringToPath, validate } from "../utils";
+import { stringToPath } from "../utils";
+import { useFieldLike } from "./use-field-like";
+import { useFieldLikeSync } from "./use-field-like-sync";
 
 export interface FieldRenderProps<T = any> extends FieldInstanceProps<T> {
   children: (props: FieldInstance<T>) => JSX.Element;
-  initialValue?: T;
-  listenTo?: string[];
 }
 
 function FieldComp<T>(
@@ -27,45 +24,28 @@ function FieldComp<T>(
 ) {
   const formContext = useContext(FormContext);
 
-  const {
-    formFieldsRef,
-    recomputeErrors,
-    recomputeIsTouched,
-    recomputeIsDirty,
-  } = formContext;
-
   const { name } = props;
 
   const _normalizedDotName = useMemo(() => {
     return stringToPath(name).join(".");
   }, [name]);
 
-  const [value, _setValue] = useState<T>(props.initialValue ?? ("" as T));
-  const valueRef = useRef(value);
-
-  valueRef.current = value;
-  const [errors, setErrors] = useState<string[]>([]);
-  const [isTouched, setIsTouched] = useState<boolean>(false);
-  const [isDirty, setIsDirty] = useState<boolean>(false);
-
-  const runFieldValidation = useCallback(
-    (validationFnName: "onChangeValidate" | "onBlurValidate", val: T) => {
-      let validationFn = props.onChangeValidate;
-      if (validationFnName === "onBlurValidate") {
-        validationFn = props.onBlurValidate;
-      }
-      if (validationFn) {
-        validate(val, formContext, validationFn)
-          .then(() => {
-            setErrors([]);
-          })
-          .catch((error: string | ZodError) => {
-            setErrors(getValidationError(error as ZodError | string));
-          });
-      }
-    },
-    [formContext, props.onChangeValidate, props.onBlurValidate]
-  );
+  const {
+    value,
+    setErrors,
+    errors,
+    setIsDirty,
+    setIsTouched,
+    setValue,
+    isTouched,
+    isDirty,
+    isValid,
+    runFieldValidation,
+    valueRef,
+  } = useFieldLike<T, FieldInstance<T>>({
+    props,
+    initialValue: "" as T,
+  });
 
   const onBlur = useCallback(() => {
     setIsTouched(true);
@@ -81,32 +61,6 @@ function FieldComp<T>(
 
     runFieldValidation("onBlurValidate", valueRef.current);
   }, []);
-
-  const setValue = useCallback(
-    (val: T) => {
-      setIsDirty(true);
-      setIsTouched(true);
-      _setValue(val);
-
-      /**
-       * Call `listenTo` field subscribers for other fields.
-       *
-       * Placed into a `setTimeout` so that the `setValue` call can finish before the `onChangeListenerRefs` are called.
-       */
-      setTimeout(() => {
-        formContext.onChangeListenerRefs.current[props.name]?.forEach((fn) =>
-          fn()
-        );
-      }, 0);
-
-      runFieldValidation("onChangeValidate", val);
-    },
-    [runFieldValidation, formContext, props.name]
-  );
-
-  const isValid = useMemo(() => {
-    return errors.length === 0;
-  }, [errors]);
 
   const fieldInstance = useMemo(() => {
     return {
@@ -124,8 +78,8 @@ function FieldComp<T>(
       _normalizedDotName,
     };
   }, [
-    value,
     props,
+    value,
     setErrors,
     errors,
     setIsDirty,
@@ -140,106 +94,15 @@ function FieldComp<T>(
 
   const mutableRef = useRef<FieldInstance<T>>(fieldInstance);
 
-  /**
-   * Add mutable ref to formFieldsRef
-   */
-  useLayoutEffect(() => {
-    mutableRef.current.props = props;
-    const newMutable = mutableRef.current;
-    formFieldsRef.current.push(newMutable);
-
-    return () => {
-      formFieldsRef.current.slice(formFieldsRef.current.indexOf(newMutable), 1);
-    };
-  }, [props]);
-
-  /**
-   * Setup `listenTo` field listeners for this field
-   */
-  useLayoutEffect(() => {
-    if (!props.listenTo || props.listenTo.length === 0) return;
-
-    function onChangeListener() {
-      runFieldValidation("onChangeValidate", valueRef.current);
-    }
-
-    function onBlurListener() {
-      runFieldValidation("onBlurValidate", valueRef.current);
-    }
-
-    function addListenerToListenToItem(
-      refTypeName: "onChangeListenerRefs" | "onBlurListenerRefs",
-      fieldName: string,
-      listener: () => void
-    ) {
-      // Make sure there's an array for the field
-      formContext[refTypeName].current[fieldName] =
-        formContext[refTypeName].current[fieldName] ?? [];
-      // Add the listener
-      formContext[refTypeName].current[fieldName].push(listener);
-      // Remove the listener
-      return () => {
-        formContext[refTypeName].current[fieldName].splice(
-          formContext[refTypeName].current[fieldName].indexOf(listener),
-          1
-        );
-      };
-    }
-
-    const fns = props.listenTo.flatMap((fieldName) => {
-      const onChangeFunctions = addListenerToListenToItem(
-        "onChangeListenerRefs",
-        fieldName,
-        onChangeListener
-      );
-      const onBlurFunctions = addListenerToListenToItem(
-        "onBlurListenerRefs",
-        fieldName,
-        onBlurListener
-      );
-      return [onChangeFunctions, onBlurFunctions];
-    });
-
-    return () => fns.forEach((fn) => fn());
-  }, [formContext, props.listenTo, runFieldValidation]);
-
-  /**
-   * Sync the values with the mutable ref
-   */
-  useLayoutEffect(() => {
-    mutableRef.current.value = value;
-  }, [value]);
-
-  useLayoutEffect(() => {
-    mutableRef.current.errors = errors;
-  }, [errors]);
-
-  useLayoutEffect(() => {
-    mutableRef.current.isDirty = isDirty;
-  }, [isDirty]);
-
-  useLayoutEffect(() => {
-    mutableRef.current.isValid = isValid;
-  }, [isValid]);
-
-  useLayoutEffect(() => {
-    mutableRef.current.isTouched = isTouched;
-  }, [isTouched]);
-
-  /**
-   * Recompute form errors when field errors change
-   */
-  useLayoutEffect(() => {
-    recomputeErrors();
-  }, [errors, recomputeErrors]);
-
-  useLayoutEffect(() => {
-    recomputeIsTouched();
-  }, [isTouched, recomputeErrors]);
-
-  useLayoutEffect(() => {
-    recomputeIsDirty();
-  }, [isDirty, recomputeErrors]);
+  useFieldLikeSync<T, FieldInstance<T>>({
+    mutableRef,
+    props,
+    value,
+    errors,
+    isValid,
+    isDirty,
+    isTouched,
+  });
 
   useImperativeHandle(ref, () => fieldInstance, [fieldInstance]);
